+++ library/src/main/android/permissions/dispatcher/DefaultChecker.java
@@ -1,0 +1,108 @@
+package permissions.dispatcher;
+
+import android.app.Activity;
+import android.content.Context;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.os.Build;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.util.SimpleArrayMap;
+
+import static android.support.v4.content.PermissionChecker.checkSelfPermission;
+
+final class DefaultChecker implements Checker {
+
+    /**
+     * Map of dangerous permissions introduced in later framework versions.
+     * Used to conditionally bypass permission-hold checks on older devices.
+     */
+    private static final SimpleArrayMap<String, Integer> MIN_SDK_PERMISSIONS;
+
+    static {
+        MIN_SDK_PERMISSIONS = new SimpleArrayMap<String, Integer>(8);
+        MIN_SDK_PERMISSIONS.put("com.android.voicemail.permission.ADD_VOICEMAIL", 14);
+        MIN_SDK_PERMISSIONS.put("android.permission.BODY_SENSORS", 20);
+        MIN_SDK_PERMISSIONS.put("android.permission.READ_CALL_LOG", 16);
+        MIN_SDK_PERMISSIONS.put("android.permission.READ_EXTERNAL_STORAGE", 16);
+        MIN_SDK_PERMISSIONS.put("android.permission.USE_SIP", 9);
+        MIN_SDK_PERMISSIONS.put("android.permission.WRITE_CALL_LOG", 16);
+        MIN_SDK_PERMISSIONS.put("android.permission.SYSTEM_ALERT_WINDOW", 23);
+        MIN_SDK_PERMISSIONS.put("android.permission.WRITE_SETTINGS", 23);
+    }
+
+    private static volatile int targetSdkVersion = -1;
+
+    @Override
+    public boolean verifyPermissions(int... grantResults) {
+        if (grantResults.length == 0) {
+            return false;
+        }
+        for (int result : grantResults) {
+            if (result != PackageManager.PERMISSION_GRANTED) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public boolean hasSelfPermissions(Context context, String... permissions) {
+        for (String permission : permissions) {
+            if (permissionExists(permission) && !hasSelfPermission(context, permission)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public boolean shouldShowRequestPermissionRationale(Activity activity, String... permissions) {
+        for (String permission : permissions) {
+            if (ActivityCompat.shouldShowRequestPermissionRationale(activity, permission)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public boolean isDeniedUnderMarshmallow(Context context, String... permissions) {
+        return getTargetSdkVersion(context) < 23 && !hasSelfPermissions(context, permissions);
+    }
+
+    private static int getTargetSdkVersion(Context context) {
+        if (targetSdkVersion != -1) {
+            return targetSdkVersion;
+        }
+        try {
+            PackageInfo packageInfo = context.getPackageManager().getPackageInfo(context.getPackageName(), 0);
+            targetSdkVersion = packageInfo.applicationInfo.targetSdkVersion;
+        } catch (PackageManager.NameNotFoundException ignored) {
+        }
+        return targetSdkVersion;
+    }
+
+    private static boolean permissionExists(String permission) {
+        // Check if the permission could potentially be missing on this device
+        Integer minVersion = MIN_SDK_PERMISSIONS.get(permission);
+        // If null was returned from the above call, there is no need for a device API level check for the permission;
+        // otherwise, we check if its minimum API level requirement is met
+        return minVersion == null || Build.VERSION.SDK_INT >= minVersion;
+    }
+
+    /**
+     * Determine context has access to the given permission.
+     * <p>
+     * This is a workaround for RuntimeException of Parcel#readException.
+     * For more detail, check this issue https://github.com/hotchemi/PermissionsDispatcher/issues/107
+     *
+     * @param context    context
+     * @param permission permission
+     * @return returns true if context has access to the given permission, false otherwise.
+     * @see #hasSelfPermissions(Context, String...)
+     */
+    private static boolean hasSelfPermission(Context context, String permission) {
+        try {
+            return checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED;
+        } catch (RuntimeException t) {
+            return false;
+        }
+    }
+}
