+++ library/src/main/java/permissions/dispatcher/RuntimePermissions.java
@@ -11,4 +11,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.CLASS)
 public @interface RuntimePermissions {
+    // Class<Checker> checker() default DefaultChecker.class;
 }
